skip UpdateConnectionPairPort if podman is not active
when starting a VM if the ssh port chosen is already in use, it tries to reassign it.
During this process it also fixes the connections saved previously and their ports.
However if podman is not in use, this step should be skipped.
E.g. when using macadam there is no concept of connections

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>