generate user-data config file based on settings on hyperv
if users provide their custom user-data we use theirc file. However,
user-mode networking requires some customization on the VM and we cannot
expect the user to provide such values. So, if users enable user-mode
networking and they provide a custom cloud-init, we update it to apply
the settings needed for user-mode networking to work fine.

if users provide a user-data file and they rely on system networking, we
keep their user-data as is.

If users do not provide any cloud init configuration, it creates a
default one.

Signed-off-by: lstocchi <lstocchi@redhat.com>